pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf4vllm-mcp"
version = "1.0.0"
description = "Block-based PDF extraction MCP server optimized for LLM consumption"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PDF MCP for vLLM Contributors"}
]
keywords = ["mcp", "pdf", "llm", "extraction", "claude", "vision", "ocr", "vllm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Filters",
]

dependencies = [
    "mcp>=0.9.0",
    "pypdfium2>=4.0.0",
    "pikepdf>=8.0.0",
    "pdfplumber>=0.10.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pdfminer.six>=20221105",
]

[project.optional-dependencies]
test = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/PyJudge/pdf4vllm-mcp"
Documentation = "https://github.com/PyJudge/pdf4vllm-mcp#readme"
Repository = "https://github.com/PyJudge/pdf4vllm-mcp"
Issues = "https://github.com/PyJudge/pdf4vllm-mcp/issues"

[project.scripts]
pdf4vllm-test = "test_server:run"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["*.py"]
